SliceSampling

Documentation for SliceSampling.

SliceSampling.LatentSlice

SliceSampling.Slice

SliceSampling.SliceDoublingOut

SliceSampling.SliceSteppingOut

SliceSampling.initial_sample

SliceSampling.LatentSlice — Type

LatentSlice(beta)

Latent slice sampling algorithm by Li and Walker[LW2023].

Fields

beta::Real: Beta parameter of the Gamma distribution of the auxiliary variables.

SliceSampling.Slice — Type

Slice(window)

Univariate slice sampling with a fixed initial interval (Scheme 2 by Neal[N2003])

Fields

window::Union{<:Real, <:AbstractVector}: Proposal window.

SliceSampling.SliceDoublingOut — Type

SliceDoublingOut(max_doubling_out, window)
SliceDoublingOut(window)

Univariate slice sampling by automatically adapting the initial interval through the "doubling-out" procedure (Scheme 4 by Neal[N2003])

Fields

max_doubling_out: Maximum number of "doubling outs" (default: 8).

window::Union{<:Real, <:AbstractVector}: Proposal window.

SliceSampling.SliceSteppingOut — Type

SliceSteppingOut(max_stepping_out, window)
SliceSteppingOut(window)

Univariate slice sampling by automatically adapting the initial interval through the "stepping-out" procedure (Scheme 3 by Neal[N2003])

Fields

max_stepping_out: Maximum number of "stepping outs" (default: 32).

window::Union{<:Real, <:AbstractVector}: Proposal window.

SliceSampling.initial_sample — Function

initial_sample(rng, model)

Return the initial sample for the model using the random number generator rng.

Arguments

rng::Random.AbstractRNG: Random number generator.

model: the model of interest.