Latent Slice Sampling

Introduction

Latent slice sampling is a recent vector-valued slice sampling algorithm proposed by Li and Walker[LW2023]. Unlike other slice sampling algorithms, it treats the "search intervals" as auxiliary variables and adapts them along the samples from the log-target in a Gibbs-type scheme.

Description

Specifically, the extended joint density of the latent slice sampler is as follows:

\[ p(y, s, l) = \pi(y) \, p(s) \, \operatorname{Uniform}\left(l; \; y - s/2,\, y + s/2\right),\]

where $y$ is the parameters of the log-target $\pi$, $s$ is the width of the search interval and $l$ is the centering of the search interval relative to $y$. Naturally, the sampler operates as a blocked-Gibbs sampler

\[\begin{aligned} l_n &\sim \operatorname{Uniform}\left(l; \; y_{n-1} - s_{n-1}/2,\, y_{n-1} + s_{n-1}/2\right) \\ s_n &\sim p(s \mid y_{n-1}, l_{n}) \\ y_n &\sim \operatorname{shrinkage}\left(y \mid s_n, l_n\right), \end{aligned}\]

where $y$ is updated using the shrinkage procedure by Neal[N2003] using the initial interval formed by $(s, l)$. Therefore, the latent slice sampler can be regarded as an automatic tuning mechanism of the "initial interval" of slice samplers.

The only tunable parameter of the algorithm is then the distribution of the width $p(s)$. For this, Li and Walker recommend

\[ p(s; \beta) = \operatorname{Gamma}(s; 2, \beta),\]

where $\beta$ is a tunable parameter. The use of the gamma distribution is somewhat important since the complete conditional $p(s \mid y, l)$ needs to be available in closed-form for efficiency. (It is a truncated exponential distribution in case of the gamma.) Therefore, we only provide control over $\beta$.

Info

Since the search interval variables are states of the Markov chain, this sampler is not reversible with respect to the samples of the log-target $y$.

Interface

SliceSampling.LatentSlice — Type

LatentSlice(beta)

Latent slice sampling algorithm by Li and Walker[LW2023].

Arguments

beta::Real: Beta parameter of the Gamma distribution of the auxiliary variables.

Keyword Arguments

max_proposals::Int: Maximum number of proposals allowed until throwing an error (default: typemax(Int)).

LW2023Li, Y., & Walker, S. G. (2023). A latent slice sampling algorithm. Computational Statistics & Data Analysis, 179, 107652.

N2003Neal, R. M. (2003). Slice sampling. The annals of statistics, 31(3), 705-767.